# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/agent-office/profiles/{profileId}": {
      "put": {
        "operationId": "AgentOfficeController_updateProfile",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      },
      "delete": {
        "operationId": "AgentOfficeController_fireAgent",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
    "/api/user/agent-office/profiles/{profileId}/station": {
      "post": {
        "operationId": "AgentOfficeController_assignStation",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
    "/api/user/agent-office/profiles/{profileId}/use": {
      "post": {
        "operationId": "AgentOfficeController_directUse",
        "parameters": [
          {
            "name": "profileId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
    "/api/user/agent-office/tasks": {
      "get": {
        "operationId": "AgentOfficeController_getTasks",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      },
      "post": {
        "operationId": "AgentOfficeController_createTask",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
    "/api/user/agent-office/tasks/{taskId}": {
      "get": {
        "operationId": "AgentOfficeController_getTask",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
    "/api/user/agent-office/tasks/{taskId}/urgent": {
      "post": {
        "operationId": "AgentOfficeController_markUrgent",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
    "/api/user/agent-office/tasks/{taskId}/skip": {
      "post": {
        "operationId": "AgentOfficeController_skipTask",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
...
    "/api/user/agent-office/tasks/{taskId}/cancel": {
      "post": {
        "operationId": "AgentOfficeController_cancelTask",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
    "/api/user/agent-office/tasks/{taskId}/retry": {
      "post": {
        "operationId": "AgentOfficeController_retryTask",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
...
    "/api/user/agent-office/history": {
      "get": {
        "operationId": "AgentOfficeController_getHistory",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOffice"
        ]
      }
    },
...
    "/api/v1/agent-office/profiles/{id}": {
      "put": {
        "operationId": "AgentOfficeV1Controller_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      },
      "delete": {
        "operationId": "AgentOfficeV1Controller_removeProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/profiles/{id}/use": {
      "post": {
        "operationId": "AgentOfficeV1Controller_use",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/profiles/{id}/station": {
      "post": {
        "operationId": "AgentOfficeV1Controller_station",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/tasks": {
      "get": {
        "operationId": "AgentOfficeV1Controller_listTasks",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      },
      "post": {
        "operationId": "AgentOfficeV1Controller_createTask",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/history": {
      "get": {
        "operationId": "AgentOfficeV1Controller_history",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/tasks/{id}": {
      "get": {
        "operationId": "AgentOfficeV1Controller_getTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      },
      "delete": {
        "operationId": "AgentOfficeV1Controller_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/tasks/{id}/urgent": {
      "post": {
        "operationId": "AgentOfficeV1Controller_urgent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/tasks/{id}/skip": {
      "post": {
        "operationId": "AgentOfficeV1Controller_skip",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/tasks/reorder": {
      "post": {
        "operationId": "AgentOfficeV1Controller_reorder",
        "parameters": [
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/tasks/{id}/cancel": {
      "post": {
        "operationId": "AgentOfficeV1Controller_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
    "/api/v1/agent-office/tasks/{id}/retry": {
      "post": {
        "operationId": "AgentOfficeV1Controller_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AgentOfficeV1"
        ]
      }
    },
...
  },
  "info": {
    "title": "ZhiPath Platform API",
    "description": "智途 ZhiPath、CodeNova 与后续定制客户端共享平台 API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "client-app": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Client-App"
      }
    },
...
  }
}